+++ travel-planner/src/main.tsx
@@ -1,17 +1,14 @@
+// src/main.tsx
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
-import Layout from "./components/layout/Layout.tsx";
-import AppRoutes from "./routes/index.tsx";
+import App from "./App.tsx";
+import { BrowserRouter as Router } from "react-router-dom";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <div className="flex flex-col min-h-screen">
-      <Layout>
-        <main className="grow pt-20">
-          <AppRoutes />
-        </main>
-      </Layout>
-    </div>
+    <Router>
+      <App />
+    </Router>
   </StrictMode>
 );